requirements.txt
fastapi
uvicorn
oracledb
python-dotenv
pydantic

langchain
langchain_pinecone
langchain_community
pinecone
sentence-transformers

